refactor bundle adjust to functions
